move wizard to subfolder
